refactor(gui): split oversized state and client modules
