+++ gui/src/app/sessionStore.ts
@@ -1,652 +1,29 @@
 import { createStore } from "zustand/vanilla";
 
-import * as desktopClient from "../services/desktop/client";
 import type { Attachment } from "@/components/attachments/attachmentTypes";
+import * as desktopClient from "../services/desktop/client";
 import type {
   ChatBinding,
-  ConversationSessionSummary,
   ConversationViewSnapshot,
   PromptSettings,
   RuntimeStatus,
-  SessionSnapshot,
-  WorkspaceSession,
 } from "../services/desktop/types/contracts";
-import type { WorkspaceBoardSelection } from "./types/sessionContext";
-import type {
-  PromptDraftEntry,
-  SessionStoreSetter,
-  SessionStoreState,
-  WorkspaceMetaState,
-} from "./types/sessionStore";
-
-const GLOBAL_WORKSPACE_META_KEY = "__global__";
-
-export function getConversationStoreKey(
-  workspacePath: string,
-  conversationId: string,
-): string {
-  return `${workspacePath}::${conversationId}`;
-}
-
-export function getConversationStoreKeyForBinding(binding: ChatBinding): string {
-  return getConversationStoreKey(binding.workspacePath, binding.conversationId);
-}
-
-export function getWorkspaceMetaStoreKey(workspacePath: string | null): string {
-  return workspacePath ?? GLOBAL_WORKSPACE_META_KEY;
-}
-
-export function areChatBindingsEqual(
-  left: ChatBinding | null,
-  right: ChatBinding | null,
-): boolean {
-  if (left === right) {
-    return true;
-  }
-
-  if (left == null || right == null) {
-    return false;
-  }
-
-  return (
-    left.workspacePath === right.workspacePath &&
-    left.conversationId === right.conversationId
-  );
-}
-
-function getConversationDraftEntry(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-): PromptDraftEntry | null {
-  return drafts[key] ?? null;
-}
-
-function writePromptDraftEntry(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-  entry: PromptDraftEntry | null,
-): Record<string, PromptDraftEntry> {
-  const current = drafts[key] ?? null;
-  const nextEntry =
-    entry == null ||
-    (entry.value === "" &&
-      !entry.isPending &&
-      !entry.isPlanningMode &&
-      !entry.isUltraMode)
-      ? null
-      : entry;
-
-  if (nextEntry == null) {
-    if (current == null) {
-      return drafts;
-    }
-
-    const nextDrafts = { ...drafts };
-    delete nextDrafts[key];
-    return nextDrafts;
-  }
-
-  if (
-    current?.value === nextEntry.value &&
-    current?.isPending === nextEntry.isPending &&
-    current?.isPlanningMode === nextEntry.isPlanningMode &&
-    current?.isUltraMode === nextEntry.isUltraMode
-  ) {
-    return drafts;
-  }
-
-  return {
-    ...drafts,
-    [key]: nextEntry,
-  };
-}
-
-function setPromptDraftEntryValue(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-  value: string,
-): Record<string, PromptDraftEntry> {
-  const current = getConversationDraftEntry(drafts, key) ?? {
-    isPending: false,
-    isPlanningMode: false,
-    isUltraMode: false,
-    value: "",
-  };
-
-  return writePromptDraftEntry(drafts, key, { ...current, value });
-}
-
-function setPromptDraftEntryPending(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-  isPending: boolean,
-): Record<string, PromptDraftEntry> {
-  const current = getConversationDraftEntry(drafts, key) ?? {
-    isPending: false,
-    isPlanningMode: false,
-    isUltraMode: false,
-    value: "",
-  };
-
-  return writePromptDraftEntry(drafts, key, { ...current, isPending });
-}
-
-function setPromptDraftEntryPlanningMode(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-  isPlanningMode: boolean,
-): Record<string, PromptDraftEntry> {
-  const current = getConversationDraftEntry(drafts, key) ?? {
-    isPending: false,
-    isPlanningMode: false,
-    isUltraMode: false,
-    value: "",
-  };
-
-  return writePromptDraftEntry(drafts, key, { ...current, isPlanningMode });
-}
-
-function setPromptDraftEntryUltraMode(
-  drafts: Record<string, PromptDraftEntry>,
-  key: string,
-  isUltraMode: boolean,
-): Record<string, PromptDraftEntry> {
-  const current = getConversationDraftEntry(drafts, key) ?? {
-    isPending: false,
-    isPlanningMode: false,
-    isUltraMode: false,
-    value: "",
-  };
-
-  return writePromptDraftEntry(drafts, key, { ...current, isUltraMode });
-}
-
-function movePromptDraftEntry(
-  drafts: Record<string, PromptDraftEntry>,
-  fromKey: string | null,
-  toKey: string | null,
-): Record<string, PromptDraftEntry> {
-  if (fromKey == null || toKey == null || fromKey === toKey) {
-    return drafts;
-  }
-
-  const draft = getConversationDraftEntry(drafts, fromKey);
-  if (draft == null) {
-    return drafts;
-  }
-
-  const nextDrafts = {
-    ...writePromptDraftEntry(drafts, toKey, draft),
-  };
-  delete nextDrafts[fromKey];
-  return nextDrafts;
-}
-
-export function getSelectionFromSnapshot(
-  snapshot: SessionSnapshot,
-): WorkspaceBoardSelection {
-  if (
-    snapshot.activeWorkspacePath != null &&
-    snapshot.activeConversationId != null
-  ) {
-    return {
-      kind: "single-chat",
-      chat: {
-        workspacePath: snapshot.activeWorkspacePath,
-        conversationId: snapshot.activeConversationId,
-      },
-    };
-  }
-
-  if (snapshot.activeWorkspacePath != null) {
-    return {
-      kind: "workspace-draft",
-      workspacePath: snapshot.activeWorkspacePath,
-    };
-  }
-
-  return { kind: "empty" };
-}
-
-function deriveConversationViews(snapshot: SessionSnapshot) {
-  if (snapshot.conversationViews.length > 0) {
-    return snapshot.conversationViews;
-  }
-
-  if (
-    snapshot.activeWorkspacePath != null &&
-    snapshot.activeConversationId != null
-  ) {
-    return [
-      {
-        activeRequestIds: snapshot.visibleActiveRequestIds,
-        conversationId: snapshot.activeConversationId,
-        followup: snapshot.visibleFollowup,
-        messages: snapshot.visibleMessages,
-        requestAgentIds: snapshot.visibleRequestAgentIds,
-        todos: snapshot.visibleTodos,
-        workspacePath: snapshot.activeWorkspacePath,
-      } satisfies ConversationViewSnapshot,
-    ];
-  }
-
-  return [];
-}
-
-function buildWorkspacesByPath(workspaces: WorkspaceSession[]) {
-  const byPath: Record<string, WorkspaceSession> = {};
-  for (const workspace of workspaces) {
-    byPath[workspace.workspacePath] = workspace;
-  }
-  return byPath;
-}
-
-function buildConversationSummariesByKey(workspaces: WorkspaceSession[]) {
-  const byKey: Record<string, ConversationSessionSummary> = {};
-
-  for (const workspace of workspaces) {
-    for (const conversation of workspace.conversations) {
-      byKey[
-        getConversationStoreKey(
-          workspace.workspacePath,
-          conversation.conversationId,
-        )
-      ] = conversation;
-    }
-  }
-
-  return byKey;
-}
-
-export function areSelectionsEqual(
-  left: WorkspaceBoardSelection,
-  right: WorkspaceBoardSelection,
-): boolean {
-  if (left === right) {
-    return true;
-  }
-
-  if (left.kind !== right.kind) {
-    return false;
-  }
-
-  switch (left.kind) {
-    case "empty":
-      return true;
-    case "single-chat": {
-      const rightSelection = right as Extract<
-        WorkspaceBoardSelection,
-        { kind: "single-chat" }
-      >;
-      return areChatBindingsEqual(left.chat, rightSelection.chat);
-    }
-    case "workspace-draft": {
-      const rightSelection = right as Extract<
-        WorkspaceBoardSelection,
-        { kind: "workspace-draft" }
-      >;
-      return left.workspacePath === rightSelection.workspacePath;
-    }
-    case "saved-workspace": {
-      const rightSelection = right as Extract<
-        WorkspaceBoardSelection,
-        { kind: "saved-workspace" }
-      >;
-      return (
-        left.workspace.id === rightSelection.workspace.id &&
-        left.workspace.layoutJson === rightSelection.workspace.layoutJson &&
-        left.workspace.name === rightSelection.workspace.name &&
-        left.workspace.updatedAt === rightSelection.workspace.updatedAt &&
-        areChatBindingsEqual(left.activeChat, rightSelection.activeChat)
-      );
-    }
-  }
-}
-
-function setWorkspaceMetaState(
-  currentState: SessionStoreState,
-  workspacePath: string | null,
-  updater: (currentMeta: WorkspaceMetaState) => WorkspaceMetaState,
-) {
-  const workspaceMetaKey = getWorkspaceMetaStoreKey(workspacePath);
-  const currentMeta =
-    currentState.workspaceMetaByKey[workspaceMetaKey] ?? defaultWorkspaceMetaState;
-  const nextMeta = updater(currentMeta);
-
-  if (
-    nextMeta.promptSettings === currentMeta.promptSettings &&
-    nextMeta.promptSettingsLoaded === currentMeta.promptSettingsLoaded &&
-    nextMeta.runtimeStatus === currentMeta.runtimeStatus &&
-    nextMeta.runtimeStatusLoaded === currentMeta.runtimeStatusLoaded
-  ) {
-    return currentState.workspaceMetaByKey;
-  }
-
-  return {
-    ...currentState.workspaceMetaByKey,
-    [workspaceMetaKey]: nextMeta,
-  };
-}
-
-const defaultWorkspaceMetaState: WorkspaceMetaState = {
-  promptSettings: null,
-  promptSettingsLoaded: false,
-  runtimeStatus: null,
-  runtimeStatusLoaded: false,
-};
-
-function createSessionStoreState(set: SessionStoreSetter): SessionStoreState {
-  return {
-    activeConversationId: null,
-    activeWorkspacePath: null,
-    attachmentsByKey: {},
-    conversationSummariesByKey: {},
-    conversationViewsByKey: {},
-    isBootstrapped: false,
-    isOpeningProject: false,
-    promptDraftsByKey: {},
-    requestTimingsByConversationId: {},
-    savedWorkspaces: [],
-    selection: { kind: "empty" },
-    uiError: null,
-    workspaceMetaByKey: {},
-    workspaces: [],
-    workspacesByPath: {},
-    applySessionSnapshot: (snapshot) => {
-      const nextViews = deriveConversationViews(snapshot);
-      set((current) => {
-        const now = Date.now();
-        let nextRequestTimingsByConversationId =
-          current.requestTimingsByConversationId;
-
-        if (nextViews.length > 0) {
-          let didChangeTimings = false;
-          const nextTimingsState = {
-            ...current.requestTimingsByConversationId,
-          };
-
-          for (const view of nextViews) {
-            const currentConversationTimings =
-              current.requestTimingsByConversationId[view.conversationId] ?? {};
-            const nextConversationTimings = { ...currentConversationTimings };
-            const activeRequestIdSet = new Set(view.activeRequestIds ?? []);
-
-            for (const requestId of view.activeRequestIds ?? []) {
-              if (nextConversationTimings[requestId] == null) {
-                nextConversationTimings[requestId] = {
-                  completedAtMs: null,
-                  startedAtMs: now,
-                };
-                didChangeTimings = true;
-              }
-            }
-
-            for (const [requestId, timing] of Object.entries(
-              nextConversationTimings,
-            )) {
-              if (
-                timing.completedAtMs == null &&
-                !activeRequestIdSet.has(requestId)
-              ) {
-                nextConversationTimings[requestId] = {
-                  ...timing,
-                  completedAtMs: now,
-                };
-                didChangeTimings = true;
-              }
-            }
-
-            nextTimingsState[view.conversationId] = nextConversationTimings;
-          }
-
-          if (didChangeTimings) {
-            nextRequestTimingsByConversationId = nextTimingsState;
-          }
-        }
-
-        const nextConversationViewsByKey = { ...current.conversationViewsByKey };
-        for (const view of nextViews) {
-          nextConversationViewsByKey[
-            getConversationStoreKey(view.workspacePath, view.conversationId)
-          ] = view;
-        }
-
-        return {
-          activeConversationId: snapshot.activeConversationId,
-          activeWorkspacePath: snapshot.activeWorkspacePath,
-          conversationSummariesByKey: buildConversationSummariesByKey(
-            snapshot.workspaces,
-          ),
-          conversationViewsByKey: nextConversationViewsByKey,
-          requestTimingsByConversationId: nextRequestTimingsByConversationId,
-          savedWorkspaces: snapshot.savedWorkspaces,
-          selection: (() => {
-            if (current.selection.kind === "saved-workspace") {
-              const savedWorkspaceSelection = current.selection;
-              const matchingSavedWorkspace = snapshot.savedWorkspaces.find(
-                (workspace) => workspace.id === savedWorkspaceSelection.workspace.id,
-              );
-              return matchingSavedWorkspace == null
-                ? getSelectionFromSnapshot(snapshot)
-                : {
-                    ...savedWorkspaceSelection,
-                    workspace: {
-                      ...savedWorkspaceSelection.workspace,
-                      name: matchingSavedWorkspace.name,
-                      updatedAt: matchingSavedWorkspace.updatedAt,
-                    },
-                  };
-            }
-
-            return getSelectionFromSnapshot(snapshot);
-          })(),
-          uiError: snapshot.uiError,
-          workspaces: snapshot.workspaces,
-          workspacesByPath: buildWorkspacesByPath(snapshot.workspaces),
-        };
-      });
-    },
-    addAttachments: (key, items) => {
-      if (key == null || items.length === 0) {
-        return;
-      }
-
-      set((current) => {
-        const existing = current.attachmentsByKey[key] ?? [];
-        const existingPaths = new Set(existing.map((item) => item.path));
-        const additions = items.filter((item) => !existingPaths.has(item.path));
-        if (additions.length === 0) {
-          return current;
-        }
-
-        return {
-          attachmentsByKey: {
-            ...current.attachmentsByKey,
-            [key]: [...existing, ...additions],
-          },
-        };
-      });
-    },
-    removeAttachment: (key, id) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => {
-        const existing = current.attachmentsByKey[key];
-        if (existing == null) {
-          return current;
-        }
-
-        const next = existing.filter((item) => item.id !== id);
-        const nextByKey = { ...current.attachmentsByKey };
-        if (next.length === 0) {
-          delete nextByKey[key];
-        } else {
-          nextByKey[key] = next;
-        }
-
-        return { attachmentsByKey: nextByKey };
-      });
-    },
-    clearAttachments: (key) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => {
-        if (current.attachmentsByKey[key] == null) {
-          return current;
-        }
-
-        const nextByKey = { ...current.attachmentsByKey };
-        delete nextByKey[key];
-        return { attachmentsByKey: nextByKey };
-      });
-    },
-    clearPromptDraft: (key) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => ({
-        promptDraftsByKey: setPromptDraftEntryValue(
-          current.promptDraftsByKey,
-          key,
-          "",
-        ),
-      }));
-    },
-    movePromptDraft: (fromKey, toKey) => {
-      set((current) => ({
-        promptDraftsByKey: movePromptDraftEntry(
-          current.promptDraftsByKey,
-          fromKey,
-          toKey,
-        ),
-      }));
-    },
-    moveAttachments: (fromKey, toKey) => {
-      if (fromKey == null || toKey == null || fromKey === toKey) {
-        return;
-      }
-
-      set((current) => {
-        const moving = current.attachmentsByKey[fromKey];
-        if (moving == null) {
-          return current;
-        }
-
-        const nextByKey = { ...current.attachmentsByKey };
-        delete nextByKey[fromKey];
-
-        const existing = nextByKey[toKey] ?? [];
-        const existingPaths = new Set(existing.map((item) => item.path));
-        nextByKey[toKey] = [
-          ...existing,
-          ...moving.filter((item) => !existingPaths.has(item.path)),
-        ];
-
-        return { attachmentsByKey: nextByKey };
-      });
-    },
-    setBoardSelection: (selection) => {
-      set((current) =>
-        areSelectionsEqual(current.selection, selection) ? current : { selection },
-      );
-    },
-    setIsBootstrapped: (value) => {
-      set((current) =>
-        current.isBootstrapped === value ? current : { isBootstrapped: value },
-      );
-    },
-    setIsOpeningProject: (value) => {
-      set((current) =>
-        current.isOpeningProject === value
-          ? current
-          : { isOpeningProject: value },
-      );
-    },
-    setPromptDraftPending: (key, isPending) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => ({
-        promptDraftsByKey: setPromptDraftEntryPending(
-          current.promptDraftsByKey,
-          key,
-          isPending,
-        ),
-      }));
-    },
-    setPromptDraftPlanningMode: (key, isPlanningMode) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => ({
-        promptDraftsByKey: setPromptDraftEntryPlanningMode(
-          current.promptDraftsByKey,
-          key,
-          isPlanningMode,
-        ),
-      }));
-    },
-    setPromptDraftUltraMode: (key, isUltraMode) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => ({
-        promptDraftsByKey: setPromptDraftEntryUltraMode(
-          current.promptDraftsByKey,
-          key,
-          isUltraMode,
-        ),
-      }));
-    },
-    setPromptDraftValue: (key, value) => {
-      if (key == null) {
-        return;
-      }
-
-      set((current) => ({
-        promptDraftsByKey: setPromptDraftEntryValue(
-          current.promptDraftsByKey,
-          key,
-          value,
-        ),
-      }));
-    },
-    setWorkspacePromptSettings: (workspacePath, promptSettings) => {
-      set((current) => ({
-        workspaceMetaByKey: setWorkspaceMetaState(
-          current,
-          workspacePath,
-          (currentMeta) => ({
-            ...currentMeta,
-            promptSettings,
-            promptSettingsLoaded: true,
-          }),
-        ),
-      }));
-    },
-    setWorkspaceRuntimeStatus: (workspacePath, runtimeStatus) => {
-      set((current) => ({
-        workspaceMetaByKey: setWorkspaceMetaState(
-          current,
-          workspacePath,
-          (currentMeta) => ({
-            ...currentMeta,
-            runtimeStatus,
-            runtimeStatusLoaded: true,
-          }),
-        ),
-      }));
-    },
-  };
-}
+import {
+  defaultWorkspaceMetaState,
+  getConversationStoreKeyForBinding,
+  getWorkspaceMetaStoreKey,
+} from "./sessionStoreHelpers";
+import { createSessionStoreState } from "./sessionStoreState";
+import type { SessionStoreState } from "./types/sessionStore";
+
+export {
+  areChatBindingsEqual,
+  areSelectionsEqual,
+  getConversationStoreKey,
+  getConversationStoreKeyForBinding,
+  getSelectionFromSnapshot,
+  getWorkspaceMetaStoreKey,
+} from "./sessionStoreHelpers";
 
 export const sessionStore = createStore<SessionStoreState>((set) =>
   createSessionStoreState(set),
@@ -772,7 +149,9 @@ export function getUiActiveWorkspacePath(
 ): string | null {
   switch (state.selection.kind) {
     case "saved-workspace":
-      return state.selection.activeChat?.workspacePath ?? state.activeWorkspacePath;
+      return (
+        state.selection.activeChat?.workspacePath ?? state.activeWorkspacePath
+      );
     case "single-chat":
       return state.selection.chat.workspacePath;
     case "workspace-draft":
@@ -897,12 +276,14 @@ export async function ensureWorkspacePromptSettingsLoaded(
 
   const request = (async () => {
     try {
-      const promptSettings = await desktopClient.getPromptSettings(workspacePath);
+      const promptSettings =
+        await desktopClient.getPromptSettings(workspacePath);
       sessionStore
         .getState()
         .setWorkspacePromptSettings(workspacePath, promptSettings);
       try {
-        const runtimeStatus = await desktopClient.getRuntimeStatus(workspacePath);
+        const runtimeStatus =
+          await desktopClient.getRuntimeStatus(workspacePath);
         sessionStore
           .getState()
           .setWorkspaceRuntimeStatus(workspacePath, runtimeStatus);
